+++ src/components/shared/sidebar/index.ts
@@ -1,3 +1,6 @@
 export { ProjectTabToolbarSidebar } from './ProjectTabToolbarSidebar';
 export { SidebarButtonGroup } from './SidebarButtonGroup';
 export { SidebarButton } from './SidebarButton';
+export { SidebarDivider } from './SidebarDivider';
+export { SidebarFieldGroup } from './SidebarFieldGroup';
+export { SidebarSectionHeader } from './SidebarSectionHeader';